requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security
Werkzeug==3.0.1
bleach==6.1.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0

# Production server
gunicorn==21.2.0